requirements.txt
requests
python-dotenv
pypdf
reportlab
tqdm
pdf2image
pillow
